+++ lib/options-fundraising.php
@@ -68,42 +68,82 @@ function nm_register_fundraising_options_metabox() {
     'description' => 'Unique code for URL. Normally a random 6 character string.',
     'id'   => 'url_code',
     'type' => 'text_small',
+    'attributes' => array(
+      'data-validation' => 'true',
+      'data-validation-required' => 'true'
+    )
+  ) );
+
+  $main_options->add_group_field( $support_section_group, array(
+    'name' => 'Show first',
+    'description' => 'Show regular or one off option first?',
+    'id'   => 'show_first',
+    'type' => 'radio_inline',
+    'options' => array(
+		  'regular' => __( 'Regular', 'cmb2' ),
+		  'oneoff'   => __( 'One off', 'cmb2' ),
+    ),
+    'default' => 'regular',
   ) );
 
   $main_options->add_group_field( $support_section_group, array(
     'name' => 'Default low regular donation',
     'id'   => 'regular_low',
     'type' => 'text_small',
+    'attributes' => array(
+      'data-validation' => 'true',
+      'data-validation-required' => 'true'
+    )
   ) );
 
   $main_options->add_group_field( $support_section_group, array(
     'name' => 'Default medium regular donation',
     'id'   => 'regular_medium',
     'type' => 'text_small',
+    'attributes' => array(
+      'data-validation' => 'true',
+      'data-validation-required' => 'true'
+    )
   ) );
 
   $main_options->add_group_field( $support_section_group, array(
     'name' => 'Default high regular donation',
     'id'   => 'regular_high',
     'type' => 'text_small',
+    'attributes' => array(
+      'data-validation' => 'true',
+      'data-validation-required' => 'true'
+    )
   ) );
 
   $main_options->add_group_field( $support_section_group, array(
     'name' => 'Default low one-off donation',
     'id'   => 'oneoff_low',
     'type' => 'text_small',
+    'attributes' => array(
+      'data-validation' => 'true',
+      'data-validation-required' => 'true'
+    )
   ) );
 
   $main_options->add_group_field( $support_section_group, array(
     'name' => 'Default medium one-off donation',
     'id'   => 'oneoff_medium',
     'type' => 'text_small',
+    'attributes' => array(
+      'data-validation' => 'true',
+      'data-validation-required' => 'true'
+    )
   ) );
 
   $main_options->add_group_field( $support_section_group, array(
     'name' => 'Default high one-off donation',
     'id'   => 'oneoff_high',
     'type' => 'text_small',
+    'attributes' => array(
+      'data-validation' => 'true',
+      'data-validation-required' => 'true'
+    )
   ) );
 
   $main_options->add_field( array(